requirements.txt
psutil
numpy
matplotlib
pillow
requests
speedtest-cli
torch
torchvision
torchaudio